Update href navigation for both NetRunner & Plugins
Routing internally within a plugin can be done as usual, but routing to core components with now require prefixing any link with / to indicate root path.

As you can imagine, we had to modify <base>. Not the prettiest solution, but the lightest one that does not require systematic URL rewrite within components.